+++ src/base/DecodersAndSanitizers/Protocols/ITB/curve_and_convex/CurveNoConfigDecoderAndSanitizer.sol
@@ -2,11 +2,19 @@
 pragma solidity ^0.8.0;
 
 abstract contract CurveNoConfigDecoderAndSanitizer {
-    function addLiquidityAllCoinsAndStake(address _pool, uint[] memory, address _gauge, uint) external pure returns (bytes memory addressesFound) {
+    function addLiquidityAllCoinsAndStake(address _pool, uint256[] memory, address _gauge, uint256)
+        external
+        pure
+        returns (bytes memory addressesFound)
+    {
         addressesFound = abi.encodePacked(_pool, _gauge);
     }
 
-    function unstakeAndRemoveLiquidityAllCoins(address _pool, uint, address _gauge, uint[] memory) external pure returns (bytes memory addressesFound) {
+    function unstakeAndRemoveLiquidityAllCoins(address _pool, uint256, address _gauge, uint256[] memory)
+        external
+        pure
+        returns (bytes memory addressesFound)
+    {
         addressesFound = abi.encodePacked(_pool, _gauge);
     }
 }
