+++ src/base/DecodersAndSanitizers/Protocols/ITB/gearbox/GearboxDecoderAndSanitizer.sol
@@ -1,25 +1,25 @@
 /* SPDX-License-Identifier: UNLICENSED */
 pragma solidity ^0.8.0;
 
-import '../common/ITBContractDecoderAndSanitizer.sol';
+import "../common/ITBContractDecoderAndSanitizer.sol";
 
 abstract contract GearboxDecoderAndSanitizer is ITBContractDecoderAndSanitizer {
-    function deposit(uint, uint) external pure returns (bytes memory addressesFound) {
+    function deposit(uint256, uint256) external pure returns (bytes memory addressesFound) {
         // Nothing to sanitize or return
         return addressesFound;
     }
 
-    function withdrawSupply(uint, uint) external pure returns (bytes memory addressesFound) {
+    function withdrawSupply(uint256, uint256) external pure returns (bytes memory addressesFound) {
         // Nothing to sanitize or return
         return addressesFound;
     }
 
-    function stake(uint) external pure returns (bytes memory addressesFound) {
+    function stake(uint256) external pure returns (bytes memory addressesFound) {
         // Nothing to sanitize or return
         return addressesFound;
     }
 
-    function unstake(uint) external pure returns (bytes memory addressesFound) {
+    function unstake(uint256) external pure returns (bytes memory addressesFound) {
         // Nothing to sanitize or return
         return addressesFound;
     }
